Internal pkg (#188)
create  github.com/komuw/ong/internal/clientip
