+++ cookie/cookie.go
@@ -4,7 +4,6 @@ package cookie
 import (
 	"errors"
 	"fmt"
-	"net"
 	"net/http"
 	"strconv"
 	"strings"
@@ -12,6 +11,7 @@ import (
 	"time"
 
 	"github.com/komuw/ong/cry"
+	"github.com/komuw/ong/internal/clientip"
 )
 
 const (
@@ -118,20 +118,15 @@ func SetEncrypted(
 		enc = cry.New(secretKey)
 	})
 
-	// TODO: fix this. If `RemoteAddr` does not have a port, `SplitHostPort` complains.
-	//
-	// Note: client IP can be spoofed easily and this could lead to issues with their cookies.
-	ip, _, err := net.SplitHostPort(r.RemoteAddr)
-	if err != nil {
-		return
-	}
-
+	ip := clientip.Get(
+		// Note: client IP can be spoofed easily and this could lead to issues with their cookies.
+		r,
+	)
 	expires := strconv.Itoa(
 		int(
 			time.Now().UTC().Add(mAge).Unix(),
 		),
 	)
-
 	combined := ip + expires + value
 	encryptedEncodedVal := fmt.Sprintf(
 		"%d%s%d%s%s",
@@ -192,17 +187,12 @@ func GetEncrypted(
 		decryptedVal[lenOfIp+lenOfExpires:]
 
 	{
-		// TODO: fix this. If `RemoteAddr` does not have a port, `SplitHostPort` complains.
-		//
-		// Note: client IP can be spoofed easily and this could lead to issues with their cookies.
-		//
 		// Try and prevent replay attacks.
 		// This does not completely stop them, but it is better than nothing.
-		incomingIP, _, errS := net.SplitHostPort(r.RemoteAddr)
-		if errS != nil {
-			return nil, errS
-		}
-
+		incomingIP := clientip.Get(
+			// Note: client IP can be spoofed easily and this could lead to issues with their cookies.
+			r,
+		)
 		if ip != incomingIP {
 			return nil, errors.New("ong/cookie: mismatched IP addresses")
 		}
